backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
pydantic==2.9.2
pypdf>=4.0.0
groq>=0.4.0
openpyxl>=3.1.0
Pillow>=10.0.0
numpy>=1.24.0
easyocr>=1.7.0
